backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.36
aiomysql>=0.2.0
pymysql>=1.1.0
cryptography>=44.0.0

# Redis
redis>=5.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# Agentic Data Scientist (Core Engine)
# Installed from Gitee in Dockerfile: git+https://gitee.com/foursking1/agentic-data-scientist.git

# HTTP Client
httpx>=0.27.2
aiohttp>=3.11.0

# Testing
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0

# Development
ruff>=0.8.0
black>=24.10.0
